fix(console): move agent-state mock fixture out of tests/
